requirements.txt
flask
flask-sqlalchemy
flask-login
tensorflow
scikit-learn
nltk
pandas
numpy
joblib
google-auth-oauthlib
google-api-python-client
python-dotenv